skills/arangodb-manager/requirements.txt
mcp
python-arango
